// src/orval/type.schemas.ts
export type GetApiListingParams = {
SearchPhrase?: string;
MachineTypeIds?: string[];
MachineMakeIds?: string[];
LowestPrice?: number;
HighestPrice?: number;
Location?: StatesAndTerritoriesOfMalaysia;
Year?: number;
ListingType?: ListingType;
ListingCondition?: ListingCondition;
ListingStatus?: ListingStatus;
PageSize?: number;
PageNumber?: number;
SortBy?: string;
SortDirection?: ListSortDirection;
DistributorProfileId?: string;
};
// ...
export type StatesAndTerritoriesOfMalaysia = typeof StatesAndTerritoriesOfMalaysia[keyof typeof StatesAndTerritoriesOfMalaysia];



export const StatesAndTerritoriesOfMalaysia = {
  Default: 'Default',
  Johor: 'Johor',
  Kedah: 'Kedah',
  Kelantan: 'Kelantan',
  Malacca: 'Malacca',
  NegeriSembilan: 'NegeriSembilan',
  Pahang: 'Pahang',
  Penang: 'Penang',
  Perak: 'Perak',
  Perlis: 'Perlis',
  Sabah: 'Sabah',
  Sarawak: 'Sarawak',
  Selangor: 'Selangor',
  Terengganu: 'Terengganu',
  KualaLumpur: 'KualaLumpur',
  Labuan: 'Labuan',
  Putrajaya: 'Putrajaya',
} as const;
// ...
export interface MachineTypeDto {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  typeIconUrl?: string | null;
}
// ...
export interface MachineMakeDto {
  id?: string;
  /** @nullable */
  logoImageUrl?: string | null;
  /** @nullable */
  name?: string | null;
}
// ...
export type ListingType = typeof ListingType[keyof typeof ListingType];



export const ListingType = {
  Default: 'Default',
  Sell: 'Sell',
  Rent: 'Rent',
} as const;

export interface ListingTagDto {
  id?: string;
  listingId?: string;
  tagId?: string;
}

export type ListingStatus = typeof ListingStatus[keyof typeof ListingStatus];



export const ListingStatus = {
  Default: 'Default',
  Draft: 'Draft',
  Published: 'Published',
  Sold: 'Sold',
  Unavailable: 'Unavailable',
} as const;
// ...
export type ListingCondition = typeof ListingCondition[keyof typeof ListingCondition];



export const ListingCondition = {
  Default: 'Default',
  New: 'New',
  Used: 'Used',
  Recon: 'Recon',
} as const;
// ...
export type ListSortDirection = typeof ListSortDirection[keyof typeof ListSortDirection];



export const ListSortDirection = {
  Ascending: 'Ascending',
  Descending: 'Descending',
} as const;
// ...
export interface FilterDto {
  /** @nullable */
  machineMakes?: MachineMakeDto[] | null;
  /** @nullable */
  machineTypes?: MachineTypeDto[] | null;
  totalMachineMakes?: number;
  totalMachineTypes?: number;
}